+++ CHANGELOG.md
@@ -4,6 +4,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## Unreleased - [2024-08-20]
+
+### `Fixed`
+
+- Added error handling to `getpass.getuser()` when to prevent throwing an error when there is no user UID such as when running the container in docker. [PR 37](https://github.com/phac-nml/locidex/pull/37)
+
+- Removed print debug statement from `report.py`. [PR 37](https://github.com/phac-nml/locidex/pull/37)
 
 ## v0.2.2 - [2024-08-01]